feat/LIVE-537 hw-transport-react-native-ble package mvp
First version of the react native module that acts as a transport bridging the native ios ble transport and providing support for background running tasks such as consuming a queue of application operations (installs and uninstalls). This version also has basic error handling of things such as pairing rejection, disconnects, etc and will be further improved with the help of our lovely QA team.

Co-authored-by: Julien Elbaz <elbywan@hotmail.com>